+++ etils/epy/__init__.py
@@ -31,6 +31,7 @@ from etils.epy.py_utils import StrEnum
 from etils.epy.reraise_utils import maybe_reraise
 from etils.epy.reraise_utils import reraise
 from etils.epy.text_utils import dedent
+from etils.epy.text_utils import diff_str
 from etils.epy.text_utils import Lines
 
 # Inside tests, can use `epy.testing`